move oper data query formatter
