+++ spec/OpenDataQuery.d.ts
