+++ spec/OpenDataQueryFormatter.d.ts
@@ -1,4 +1,4 @@
-import { SqlFormatter } from "./formatter";
+import { SqlFormatter } from "../src/formatter";
 
 export declare class OpenDataQueryFormatter extends SqlFormatter {
     //
